4.22*: bloglist expansion, step10 DONE

+++ part4/blog/controllers/blogs.js
@@ -1,34 +1,67 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (req, res) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     res.json(blogs)
 })
 
-blogsRouter.post('/', async (req, res) => {
-    const blog = new Blog(req.body)
-
-    const savedBlog = await blog.save()
-    res.status(201).json(savedBlog)
+blogsRouter.get('/:id', async (req, res) => {
+    const blog = await Blog.findById(req.params.id)
+    if (blog) {
+        res.status(200).json(blog)
+    } else {
+        res.status(404).end()
+    }
 })
 
-blogsRouter.delete('/:id', async (req, res) => {
-    await Blog.findByIdAndRemove(req.params.id)
-    res.status(204).end()
-})
-
-blogsRouter.put('/:id', async (req, res) => {
+blogsRouter.post('/', userExtractor, async (req, res) => {
     const body = req.body
-    const blog = {
+    const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes
-    }
+        likes: body.likes,
+        user: req.user._id
+    })
 
-    const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, blog, { new: true })
-    res.status(200).json(updatedBlog)
+    const savedBlog = await blog.save()
+    req.user.blogs = req.user.blogs.concat(savedBlog._id)
+    await req.user.save()
+
+    res.status(201).json(savedBlog)
+})
+
+blogsRouter.delete('/:id', userExtractor, async (req, res) => {
+    const blog = await Blog.findById(req.params.id)
+
+    if ( blog.user.toString() === req.user.id.toString() ) {
+        await Blog.findByIdAndRemove(req.params.id)
+        res.status(204).end()
+    } else {
+        res.status(401).json({ error: 'Not authorized to delete this post' })
+    }
+})
+
+blogsRouter.put('/:id', userExtractor, async (req, res) => {
+    const blog = await Blog.findById(req.params.id)
+
+    if ( blog.user.toString() === req.user.id.toString() ) {
+        const body = req.body
+        const updatedBlog = {
+            title: body.title,
+            author: body.author,
+            url: body.url,
+            likes: body.likes,
+            user: req.user.id
+        }
+
+        await Blog.findByIdAndUpdate(req.params.id, updatedBlog, { new: true })
+        res.status(200).json(updatedBlog)
+    } else {
+        res.status(401).json({ error: 'Not authorized to update this post' })
+    }
 })
 
 module.exports = blogsRouter
